+++ demisto_sdk/commands/content_graph/interface/neo4j/queries/validations.py
@@ -3,17 +3,22 @@ from typing import List, Tuple
 from neo4j import Transaction, graph
 
 from demisto_sdk.commands.content_graph.common import (
+    ContentType,
     Neo4jRelationshipResult,
     RelationshipType,
 )
-from demisto_sdk.commands.content_graph.interface.neo4j.queries.common import run_query
+from demisto_sdk.commands.content_graph.interface.neo4j.queries.common import (
+    run_query,
+    versioned,
+)
 
 
 def validate_unknown_content(tx: Transaction, file_paths):
     file_paths_filter = (
-        f"WHERE content_item_from.path in {str(file_paths)}" if file_paths else ""
+        f"WHERE content_item_from.path in {file_paths}" if file_paths else ""
     )
     query = f"""
+        // Returning all the USES relationships where the target is not in the repository
         MATCH (content_item_from)-[r:{RelationshipType.USES}{{mandatorily:true}}]->(n{{not_in_repository:true}})
         {file_paths_filter}
         RETURN content_item_from, collect(r) as relationships, collect(n) as nodes_to
@@ -30,17 +35,18 @@ def validate_unknown_content(tx: Transaction, file_paths):
 
 def validate_fromversion(tx: Transaction, file_paths, from_version):
     file_paths_filter = (
-        f"AND content_item_from.path in {str(file_paths)}" if file_paths else ""
+        f"AND content_item_from.path in {file_paths}" if file_paths else ""
     )
     from_version_filter = (
-        "AND toIntegerList(split(content_item_from.fromversion, '.')) >= toIntegerList([6,5,0])"
+        f"AND {versioned('content_item_from.fromversion')} >= {versioned([6,5,0])}"
         if from_version
         else ""
     )
 
     query = f"""
+        // Returning all the USES relationships with where the target's to_version is bigger than the source's
         MATCH (content_item_from)-[r:{RelationshipType.USES}{{mandatorily:true}}]->(n)
-        WHERE toIntegerList(split(content_item_from.fromversion, ".")) < toIntegerList(split(n.fromversion, "."))
+        WHERE {versioned('content_item_from.fromversion')} < {versioned('n.fromversion')}
         {from_version_filter}
         {file_paths_filter}
         RETURN content_item_from, collect(r) as relationships, collect(n) as nodes_to
@@ -57,16 +63,17 @@ def validate_fromversion(tx: Transaction, file_paths, from_version):
 
 def validate_toversion(tx: Transaction, file_paths, to_version):
     file_paths_filter = (
-        f"AND content_item_from.path in {str(file_paths)}" if file_paths else ""
+        f"AND content_item_from.path in {file_paths}" if file_paths else ""
     )
     from_version_filter = (
-        "AND toIntegerList(split(content_item_from.toversion, '.')) >= toIntegerList([6,5,0])"
+        f"AND {versioned('content_item_from.toversion')} >= {versioned([6,5,0])}"
         if to_version
         else ""
     )
     query = f"""
+        // Returning all the USES relationships with where the target's to_version is smaller than the source's
         MATCH (content_item_from)-[r:{RelationshipType.USES}{{mandatorily: true}}]->(n)
-        WHERE toIntegerList(split(content_item_from.toversion, ".")) > toIntegerList(split(n.toversion, "."))
+        WHERE {versioned('content_item_from.toversion')} > {versioned('n.toversion')}
         {from_version_filter}
         {file_paths_filter}
         RETURN content_item_from, collect(r) as relationships, collect(n) as nodes_to
@@ -83,9 +90,10 @@ def validate_toversion(tx: Transaction, file_paths, to_version):
 
 def validate_marketplaces(tx: Transaction, file_paths):
     file_paths_filter = (
-        f"AND content_item_from.path in {str(file_paths)}" if file_paths else ""
+        f"AND content_item_from.path in {file_paths}" if file_paths else ""
     )
     query = f"""
+        // Returning all the USES relationships with where the target's marketplaces doesn't include all of the source's marketplaces
         MATCH (content_item_from)-[r:{RelationshipType.USES}{{mandatorily:true}}]->(n)
         WHERE not all(elem IN content_item_from.marketplaces WHERE elem IN n.marketplaces)
         {file_paths_filter}
@@ -101,15 +109,16 @@ def validate_marketplaces(tx: Transaction, file_paths):
     }
 
 
-def validate_duplicate_display_name(
-    tx: Transaction, pack_names: list
+def validate_multiple_packs_with_same_display_name(
+    tx: Transaction, file_paths: list
 ) -> List[Tuple[graph.Node, List[graph.Node]]]:
-    pack_names_filter = f"AND a.name in {str(pack_names)}" if pack_names else ""
+    file_paths_filter = f"AND a.path in {file_paths}" if file_paths else ""
 
     query = f"""
-        MATCH (a:Pack), (b:Pack)
+        // Returning all the packs that have the same name but different id
+        MATCH (a:{ContentType.PACK}), (b:{ContentType.PACK})
         WHERE a.name = b.name
-        {pack_names_filter}
+        {file_paths_filter}
         AND id(a) <> id(b)
         RETURN a.object_id AS a_object_id, collect(b.object_id) AS b_object_ids
         """