+++ demisto_sdk/commands/content_graph/objects/content_item.py
@@ -87,7 +87,7 @@ class ContentItem(BaseContent):
             if r.content_item == r.target
         ]
 
-    def used_by(self):
+    def used_by(self) -> List["RelationshipData"]:
         return [
             r
             for r in self.relationships_data[RelationshipType.USES]